# src/modules/board/board.controller.ts
import { inject, injectable } from "inversify";
import {
  Body,
  Controller,
  Delete,
  Get,
  Patch,
  Post,
  Query,
  Route,
  Security,
  Tags,
} from "tsoa";

import { ApiResponse } from "../../dto/ApiResponse";
import {
  IBoardCreateRequest,
  IBoardDto,
  IBoardListDto,
  IBoardUpdateRequest,
} from "./board.model";
import { BoardService } from "./board.service";

@injectable()
@Tags("Board")
@Route("api/board")
export class BoardController extends Controller {
  constructor(@inject(BoardService) private _boardService: BoardService) {
    super();
  }

  @Security("jwt")
  @Get("list")
  getBoards(
    @Query("offset") offset?: number,
    @Query("limit") limit?: number,
    @Query("projectId") projectId?: string,
  ): Promise<IBoardListDto> {
    return this._boardService
      .getBoards(offset, limit, projectId ? { projectId } : undefined)
      .then(result => ({
        offset,
        limit,
        count: result.length,
        data: result.map(res => res.toDTO()),
      }));
  }

  @Security("jwt")
  @Get("{id}")
  getBoardById(id: string): Promise<IBoardDto> {
    return this._boardService.getBoardById(id).then(res => res.toDTO());
  }

  @Security("jwt", ["role:admin"])
  @Post()
  createBoard(@Body() body: IBoardCreateRequest): Promise<IBoardDto> {
    return this._boardService.createBoard(body).then(res => res.toDTO());
  }

  @Security("jwt", ["role:admin"])
  @Patch("{id}")
  updateBoard(
    id: string,
    @Body() body: IBoardUpdateRequest,
  ): Promise<IBoardDto> {
    return this._boardService.updateBoard(id, body).then(res => res.toDTO());
  }

  @Security("jwt", ["role:admin"])
  @Delete("{id}")
  deleteBoard(id: string): Promise<string> {
    return this._boardService.deleteBoard(id);
  }

  @Security("jwt")
  @Post("{boardId}/sprints/{sprintId}/activate")
  setActiveSprint(boardId: string, sprintId: string): Promise<IBoardDto> {
    return this._boardService
      .setActiveSprint(boardId, sprintId)
      .then(res => res.toDTO());
  }
}