# swagger.json
{
	"components": {
...
		"schemas": {
...
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/ERole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string"
					},
					"challenge": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"role": {
						"$ref": "#/components/schemas/IRoleDto"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ITeamMemberDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"role": {
						"$ref": "#/components/schemas/ETeamRole"
					}
				},
				"required": [
					"id",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeamDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/ITeamMemberDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EProjectVisibility": {
				"enum": [
					"private",
					"team",
					"public"
				],
				"type": "string"
			},
			"IProjectDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/EProjectVisibility"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/ITeamDto"
						},
						"type": "array"
					},
					"lead": {
						"$ref": "#/components/schemas/IUserDto"
					}
				},
				"required": [
					"id",
					"name",
					"key",
					"visibility",
					"createdAt",
					"updatedAt",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IStatusDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"isInitial": {
						"type": "boolean"
					},
					"isFinal": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"isInitial",
					"isFinal",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EBoardType": {
				"enum": [
					"scrum",
					"kanban"
				],
				"type": "string"
			},
			"ESprintStatus": {
				"enum": [
					"future",
					"active",
					"closed"
				],
				"type": "string"
			},
			"IBoardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EBoardType"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"project": {
						"$ref": "#/components/schemas/IProjectDto"
					},
					"sprints": {
						"items": {
							"$ref": "#/components/schemas/ISprintDto"
						},
						"type": "array"
					},
					"activeSprint": {
						"$ref": "#/components/schemas/ISprintDto"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EIssueType": {
				"enum": [
					"story",
					"task",
					"bug",
					"epic",
					"subtask"
				],
				"type": "string"
			},
			"IIssueTypeDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EIssueType"
					},
					"icon": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"icon",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EPriority": {
				"enum": [
					"highest",
					"high",
					"medium",
					"low",
					"lowest"
				],
				"type": "string"
			},
			"IPriorityDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EPriority"
					},
					"icon": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"icon",
					"color",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISprintDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/ESprintStatus"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"board": {
						"$ref": "#/components/schemas/IBoardDto"
					},
					"issues": {
						"items": {
							"$ref": "#/components/schemas/IIssueDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"startDate",
					"endDate",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIssueDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"estimate": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"$ref": "#/components/schemas/IIssueTypeDto"
					},
					"priority": {
						"$ref": "#/components/schemas/IPriorityDto"
					},
					"status": {
						"$ref": "#/components/schemas/IStatusDto"
					},
					"project": {
						"$ref": "#/components/schemas/IProjectDto"
					},
					"board": {
						"$ref": "#/components/schemas/IBoardDto"
					},
					"sprint": {
						"$ref": "#/components/schemas/ISprintDto"
					},
					"assignee": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"reporter": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"parent": {
						"$ref": "#/components/schemas/IIssueDto"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/ICommentDto"
						},
						"type": "array"
					},
					"labels": {
						"items": {
							"$ref": "#/components/schemas/ILabelDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"key",
					"title",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICommentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"author": {
						"$ref": "#/components/schemas/IUserDto"
					},
					"issue": {
						"$ref": "#/components/schemas/IIssueDto"
					}
				},
				"required": [
					"id",
					"text",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILabelDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBacklogDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"isDefault": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"project": {
						"$ref": "#/components/schemas/IProjectDto"
					},
					"board": {
						"$ref": "#/components/schemas/IBoardDto"
					},
					"issues": {
						"items": {
							"$ref": "#/components/schemas/IIssueDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"isDefault",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBacklogListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IBacklogDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBacklogCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"boardId": {
						"type": "string"
					},
					"isDefault": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"projectId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBacklogUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"boardId": {
						"type": "string"
					},
					"isDefault": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IBoardListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IBoardDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBoardCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EBoardType"
					},
					"projectId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"type",
					"projectId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBoardUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EBoardType"
					},
					"projectId": {
						"type": "string"
					},
					"activeSprintId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IIssueTypeListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IIssueTypeDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIssueTypeCreateRequest": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/EIssueType"
					},
					"icon": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"icon"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIssueTypeUpdateRequest": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/EIssueType"
					},
					"icon": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/backlog/list": {
			"get": {
				"operationId": "GetBacklogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBacklogListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/backlog/{id}": {
			"get": {
				"operationId": "GetBacklogById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBacklogDto"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBacklog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBacklogDto"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBacklogUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBacklog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/backlog": {
			"post": {
				"operationId": "CreateBacklog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBacklogDto"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBacklogCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/backlog/{backlogId}/issues/{issueId}": {
			"post": {
				"operationId": "AddIssueToBacklog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBacklogDto"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "issueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveIssueFromBacklog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBacklogDto"
								}
							}
						}
					}
				},
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "issueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/board/list": {
			"get": {
				"operationId": "GetBoards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBoardListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/board/{id}": {
			"get": {
				"operationId": "GetBoardById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBoardDto"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBoardDto"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBoardUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/board": {
			"post": {
				"operationId": "CreateBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBoardDto"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBoardCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/board/{boardId}/sprints/{sprintId}/activate": {
			"post": {
				"operationId": "SetActiveSprint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBoardDto"
								}
							}
						}
					}
				},
				"tags": [
					"Board"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "boardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "sprintId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/issue-type/list": {
			"get": {
				"operationId": "GetIssueTypes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIssueTypeListDto"
								}
							}
						}
					}
				},
				"tags": [
					"IssueType"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/issue-type/{id}": {
			"get": {
				"operationId": "GetIssueTypeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIssueTypeDto"
								}
							}
						}
					}
				},
				"tags": [
					"IssueType"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateIssueType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIssueTypeDto"
								}
							}
						}
					}
				},
				"tags": [
					"IssueType"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IIssueTypeUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteIssueType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"IssueType"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/issue-type/name/{name}": {
			"get": {
				"operationId": "GetIssueTypeByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIssueTypeDto"
								}
							}
						}
					}
				},
				"tags": [
					"IssueType"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/issue-type": {
			"post": {
				"operationId": "CreateIssueType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIssueTypeDto"
								}
							}
						}
					}
				},
				"tags": [
					"IssueType"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IIssueTypeCreateRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}